Refactor the repo structure
