+++ src/test/java/com/vaadin/ui/button/demo/ButtonView.java
@@ -13,11 +13,12 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.button;
+package com.vaadin.ui.button.demo;
 
 import com.vaadin.flow.demo.DemoView;
 import com.vaadin.router.Route;
 import com.vaadin.ui.Component;
+import com.vaadin.ui.button.Button;
 import com.vaadin.ui.common.HasClickListeners.ClickEvent;
 import com.vaadin.ui.common.HtmlImport;
 import com.vaadin.ui.html.Div;
