+++ src/test/java/com/vaadin/ui/button/tests/ButtonIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.button;
+package com.vaadin.ui.button.tests;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -21,6 +21,7 @@ import org.openqa.selenium.WebElement;
 
 import com.vaadin.flow.demo.ComponentDemoTest;
 import com.vaadin.testbench.By;
+import com.vaadin.ui.button.demo.ButtonView;
 
 /**
  * Integration tests for the {@link ButtonView}.
