+++ src/test/java/com/vaadin/ui/button/tests/ButtonTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.button;
+package com.vaadin.ui.button.tests;
 
 import java.util.function.Predicate;
 
@@ -22,6 +22,7 @@ import org.junit.Test;
 
 import com.vaadin.flow.dom.Element;
 import com.vaadin.ui.Text;
+import com.vaadin.ui.button.Button;
 import com.vaadin.ui.icon.Icon;
 
 public class ButtonTest {
